# apps/landing/public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Fluxby API",
    "version": "1.0.0",
    "description": "API for managing financial transactions, categories, budgets, and analytics. This API is intended for developers building custom interfaces.",
    "license": {
      "name": "MIT",
      "url": "https://github.com/fluxby/fluxby/blob/main/LICENSE"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/api/analytics/dashboard": {
      "get": {
        "summary": "Haal dashboard statistieken op",
        "tags": [
          "Analytics"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "startDate",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Startdatum (YYYY-MM-DD)"
          },
          {
            "in": "query",
            "name": "endDate",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Einddatum (YYYY-MM-DD)"
          },
          {
            "in": "query",
            "name": "type",
            "schema": {
              "type": "string",
              "enum": [
                "income",
                "expense"
              ]
            },
            "description": "Filter op type"
          },
          {
            "in": "query",
            "name": "categoryIds",
            "schema": {
              "type": "string"
            },
            "description": "Komma-gescheiden lijst van categorie IDs"
          }
        ],
        "responses": {
          "200": {
            "description": "Dashboard statistieken inclusief totalen, maanddata en categorie verdeling"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/analytics/monthly": {
      "get": {
        "summary": "Haal maandelijkse data op voor grafieken",
        "tags": [
          "Analytics"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "startDate",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "in": "query",
            "name": "endDate",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Maandelijkse inkomsten en uitgaven"
          }
        }
      }
    },
    "/api/analytics/categories": {
      "get": {
        "summary": "Haal categorie verdeling op",
        "tags": [
          "Analytics"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "startDate",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "in": "query",
            "name": "endDate",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "in": "query",
            "name": "type",
            "schema": {
              "type": "string",
              "enum": [
                "income",
                "expense"
              ],
              "default": "expense"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Categorie verdeling met percentages"
          }
        }
      }
    },
    "/api/analytics/daily-expenses": {
      "get": {
        "summary": "Haal dagelijkse uitgaven op voor timeline",
        "tags": [
          "Analytics"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "startDate",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "required": true
          },
          {
            "in": "query",
            "name": "endDate",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Dagelijkse uitgaven inclusief dagen zonder transacties"
          }
        }
      }
    },
    "/api/analytics/balance-forecast": {
      "get": {
        "summary": "Haal saldo prognose op voor einde periode",
        "tags": [
          "Analytics"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "startDate",
            "schema": {
              "type": "string"
            },
            "description": "Start datum (YYYY-MM-DD)"
          },
          {
            "in": "query",
            "name": "endDate",
            "schema": {
              "type": "string"
            },
            "description": "Eind datum (YYYY-MM-DD)"
          }
        ],
        "responses": {
          "200": {
            "description": "Prognose met verwacht inkomen, uitgaven en eindsaldo"
          }
        }
      }
    },
...
    "/api/categories": {
      "get": {
        "summary": "Haal alle categorieën op",
        "tags": [
          "Categories"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "withCounts",
            "schema": {
              "type": "boolean"
            },
            "description": "Inclusief transactie telling per categorie"
          }
        ],
        "responses": {
          "200": {
            "description": "Lijst met categorieën"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "post": {
        "summary": "Maak een nieuwe categorie aan",
        "tags": [
          "Categories"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "parentId": {
                    "type": "integer"
                  },
                  "icon": {
                    "type": "string"
                  },
                  "color": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Categorie aangemaakt"
          },
          "400": {
            "description": "Naam is verplicht"
          }
        }
      },
      "delete": {
        "summary": "Verwijder alle categorieën",
        "tags": [
          "Categories"
        ],
        "responses": {
          "200": {
            "description": "Alle categorieën verwijderd"
          }
        }
      }
    },
...
  },
...
}